requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0 